# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarClean API",
        "description": "Documentacion oficial de la API de CarClean",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/v1/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "List sales",
                "operationId": "821a79d76738c69873078da29295f70d",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "payment_method",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cash",
                                "transfer",
                                "card"
                            ]
                        }
                    },
                    {
                        "name": "should_invoice",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create sale",
                "operationId": "a7a3f1409c7f6b7ace53809d33234ead",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "vehicle_id",
                                    "payment_method",
                                    "should_invoice",
                                    "items"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "transfer",
                                            "card"
                                        ],
                                        "example": "cash"
                                    },
                                    "paid_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "should_invoice": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "point_of_sale": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "service_id"
                                            ],
                                            "properties": {
                                                "service_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/sales/{sale}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get sale by ID",
                "operationId": "505cfba8b7a65262ee54c4ed7dd875d3",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sale retrieved successfully"
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sales"
                ],
                "summary": "Update sale",
                "operationId": "7fc7b4e1e3b1948e30dc993db8f94966",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "vehicle_id",
                                    "payment_method",
                                    "should_invoice",
                                    "items"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "transfer",
                                            "card"
                                        ],
                                        "example": "card"
                                    },
                                    "paid_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "should_invoice": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "point_of_sale": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "service_id"
                                            ],
                                            "properties": {
                                                "service_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sale updated successfully"
                    },
                    "404": {
                        "description": "Sale not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sales"
                ],
                "summary": "Delete sale",
                "operationId": "1b20bf7e6015e429f3a81eed5ff662a5",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sale deleted successfully"
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Usa: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}